새로고침 socket 문제, MemoChat 무한 리렌더링 현상 제외 하고 채팅 완성
